refactor(migration): fold simplify_test.go into per-subject test files
The migration/simplify_test.go file name encoded an ephemeral session (a
/simplify pass) instead of the code under test. Move:

- TestSafeVendorSegment* and TestRedactPassword* into migration/flyway_test.go
  (the helpers they exercise live in migration/flyway.go).
- TestSecretsProviderEmptyTenantID into migration/secrets_test.go
  (where the rest of the SecretsProvider tests live).

No semantic change. Drops one file, adds 82 lines across the two existing
per-subject test files.
